more fixes for windows build
git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta/@54802 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: 0f2d5cd62f595e9bea526220da5ed7c63dbed88f

+++ rosetta_source/src/protocols/normalmode/NormalMode.cc
@@ -324,7 +324,7 @@ NormalMode::set_harmonic_constant_map( pose::Pose const &pose ){
 			for( Size jres = 1; jres <= pose_tmp.total_residue(); ++jres ){
 				char const &SS_j = pose_tmp.secstruct( jres );
 
-				if( abs(ires - jres) < 3 || SS_j == 'L' ) continue;
+				if( abs((int)(ires - jres)) < 3 || SS_j == 'L' ) continue;
 
 				Vector const &Ocrd = pose_tmp.residue( jres ).xyz( " O  " );
 				Vector const vNO = Ncrd - Ocrd;
@@ -351,7 +351,7 @@ NormalMode::set_harmonic_constant_map( pose::Pose const &pose ){
 				k_[jatm][iatm] = k_uniform_;
 
 			} else {
-				if( abs(ires - jres) <= 1 ){
+				if( abs((int)(ires - jres)) <= 1 ){
 					k_[iatm][jatm] = k_short_;
 					k_[jatm][iatm] = k_short_;
